backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
filelock==3.20.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.12.2
itemloaders==1.3.2
jmespath==1.0.1
kombu==5.6.1
lxml==6.0.2
packaging==25.0
parsel==1.10.0
prompt_toolkit==3.0.52
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
queuelib==1.8.0
redis==7.1.0
requests==2.32.5
requests-file==3.0.1
Scrapy==2.13.4
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
tldextract==5.3.0
tomli==2.3.0
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.14
zope.interface==8.1.1
